requirements.txt
fastapi[all]
requests
pymongo>=3.12.0
pydantic>=1.8.0
starlette
httpx
pytest==8.1.2
mongomock
python-dotenv
meteostat
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0